+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Lowercase prompt in Relevance output scanner to improve quality of cosine similarity
+- Changed model used for `PromptInjection` to `JasperLS/deberta-v3-base-injection`
 
 ### Removed
 -