+++ CHANGELOG.md
@@ -19,6 +19,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 -
 
+## [0.2.2] - 2023-09-21
+
+### Fixed
+- Missing secrets detection for Github token in the final build
+
 ## [0.2.1] - 2023-09-21
 
 ### Added
@@ -142,6 +147,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - [Toxicity](./llm_guard/output_scanners/toxicity.py)
 
 [Unreleased]: https://github.com/laiyer-ai/llm-guard/commits/main
+[0.2.2]: https://github.com/laiyer-ai/llm-guard/releases/tag/v0.2.2
 [0.2.1]: https://github.com/laiyer-ai/llm-guard/releases/tag/v0.2.1
 [0.2.0]: https://github.com/laiyer-ai/llm-guard/releases/tag/v0.2.0
 [0.1.3]: https://github.com/laiyer-ai/llm-guard/releases/tag/v0.1.3